requirements.txt
numpy>=1.23
pandas>=1.5
scipy
scikit-learn>=1.1
matplotlib
seaborn
tensorflow-cpu>=2.9
pyyaml
lightgbm
tqdm
joblib
